Übersichtsseite für die Klicks der Karten angepasst.
