+++ src/BingoBundle/Resources/public/app/controller/clicks~001.js
@@ -1,4 +1,4 @@
-BingoApp.controller('BingoClicksController', function ($scope, $interval, $locale, $window, $uibModal, $log, Restangular, LoaderService, MessageService, MessageBoxService) {
+BingoApp.controller('BingoClicksController', function ($scope, $locale, $window, $uibModal, $log, Restangular, LoaderService, MessageService, MessageBoxService) {
     /**
      * Get the global logger to the local scope.
      *
@@ -72,33 +72,7 @@ BingoApp.controller('BingoClicksController', function ($scope, $interval, $local
         });
     };
 
-    //var interval = $interval($scope.getClicks, 5000);
-    $interval($scope.getClicks, 5000);
-
-    // -- LIST ORDER ---------------------------------------------------------------------------------------------------
-
-    $scope.order = 'false';
-
-    $scope.myList = [
-        {id: 0, text: 'HTML5 Boilerplate'},
-        {id: 1, text: 'AngularJS'},
-        {id: 2, text: 'Karma'},
-        {id: 3, text: 'Hello'},
-        {id: 4, text: 'World'},
-        {id: 5, text: 'How'},
-        {id: 6, text: 'Are'},
-        {id: 7, text: 'You'},
-        {id: 8, text: '?'},
-        {id: 9, text: 'I'},
-        {id: 10, text: 'write'},
-        {id: 11, text: 'more'},
-        {id: 12, text: 'to'},
-        {id: 13, text: 'make'},
-        {id: 14, text: 'the'},
-        {id: 15, text: 'list'},
-        {id: 16, text: 'longer'}
-    ];
-
+    /*
     $scope.$watch('order', function () {
         $scope.setOrder();
     });
@@ -108,22 +82,21 @@ BingoApp.controller('BingoClicksController', function ($scope, $interval, $local
 
         if ($scope.order === 'random') {
             var t = [];
-
             for (i = 0; i < $scope.myList.length; i++) {
                 var r = Math.floor(Math.random() * $scope.myList.length);
-
-                while (inArray(t, r)) {
+                while (inArray(t,r)) {
                     r = Math.floor(Math.random() * $scope.myList.length);
                 }
-
                 t.push(r);
                 $scope.myList[i].order = r;
             }
-        } else if ($scope.order === 'false') {
+        }
+        else if ($scope.order === 'false') {
             for (i = 0; i < $scope.myList.length; i++) {
                 $scope.myList[i].order = i;
             }
-        } else {
+        }
+        else {
             for (i = 0; i < $scope.myList.length; i++) {
                 $scope.myList[i].order = ($scope.myList.length - 1 - i);
             }
@@ -138,7 +111,6 @@ BingoApp.controller('BingoClicksController', function ($scope, $interval, $local
                 return true;
             }
         }
-
         return false;
     }
 
@@ -147,9 +119,9 @@ BingoApp.controller('BingoClicksController', function ($scope, $interval, $local
             var item = $scope.myList[i];
 
             // columns, left-to-right, top-to-bottom
-            var columns = 13;
-            item.column = item.order % columns;
-            item.row = Math.floor(item.order / columns);
+            var columns = 5;
+            item.column = item.order%columns;
+            item.row = Math.floor(item.order/columns);
 
             // rows, top-to-bottom, left-to-right
             // var rows = 3;
@@ -157,4 +129,5 @@ BingoApp.controller('BingoClicksController', function ($scope, $interval, $local
             // item.row = item.order%rows;
         }
     }
+    */
 });
